net: shell: Add nbr command
Move "nbr" command to a separate file.
